generate menu

+++ src/Shared/DemoMenuItem.cs
@@ -2,7 +2,7 @@
 {
     public class DemoMenuItem
     {
-        public float Order { get; set; }
+        public double Order { get; set; }
 
         public string Title { get; set; }
